backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
pydantic==2.7.0
openai==1.30.0
networkx==3.3
python-louvain==0.16
numpy==1.26.4
chromadb==0.5.0
pymupdf==1.24.3
python-dotenv==1.0.1
thefuzz==0.22.1
httpx==0.27.0
redis==5.0.4
tinyfish==0.2.4
